requirements.txt
asyncpg
sqlalchemy[asyncio]
alembic
redis
python-dotenv

python-telegram-bot[all]

pydantic
pytest

colorlog